(API/APP) Compras, fondos del usuario y transacciones
Actualizada la base de datos (precio en compras era decimal y no float)

+++ api/TiendaMovil/Models/Compra.cs
@@ -16,7 +16,7 @@ namespace TiendaMovil.Models
         [ForeignKey("PublicacionId")]
         public int PublicacionId { get; set; }
         public int Cantidad { get; set; }
-        public float Precio { get; set; }
+        public double Precio { get; set; }
         public int Estado { get; set; }
         public DateTime Creacion { get; set; }
         public Usuario Usuario { get; set; }
